+++ src/Logo.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function Logo() {
+  return (
+    <div className="logo">
+      <span role="img">🍿</span>
+      <h1>usePopcorn</h1>
+    </div>
+  );
+}
+
+export default Logo;
