+++ src/MovieList.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { useState } from "react";
+import Movie from './Movie';
+
+function MovieList({tempMovieData}) {
+    const [movies, setMovies] = useState(tempMovieData);
+
+  return (
+    <ul className="list">
+        {movies?.map((movie) => (
+         <Movie movie={movie} key={movie.imdbID} />
+        ))}
+      </ul>
+  )
+}
+
+export default MovieList
