+++ src/Navbar.jsx
@@ -1,26 +1,17 @@
 import React from 'react'
 import { useState } from 'react'
+import Search from './Search';
+import Logo from './Logo';
+import NumResults from './NumResults';
 
 
 function Navbar() {
-const [query, setQuery] = useState("");
 
   return (
     <nav className="nav-bar">
-        <div className="logo">
-          <span role="img">🍿</span>
-          <h1>usePopcorn</h1>
-        </div>
-        <input
-          className="search"
-          type="text"
-          placeholder="Search movies..."
-          value={query}
-          onChange={(e) => setQuery(e.target.value)}
-        />
-        <p className="num-results">
-          Found <strong>X</strong> results
-        </p>
+       <Logo />
+       <Search />
+       <NumResults />
       </nav>
   )
 }
