+++ src/NumResults.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function NumResults() {
+  return (
+    <div>
+      <p className="num-results">
+        Found <strong>X</strong> results
+      </p>
+    </div>
+  );
+}
+
+export default NumResults;
