+++ src/WatchedList.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import WatchedMovie from './WatchedMovie'
+
+function WatchedList({watched}) {
+  return (
+    <ul className="list">
+    {watched.map((movie) => (
+     <WatchedMovie movie={movie} key={movie.imdbID} />
+    ))}
+  </ul>
+  )
+}
+
+export default WatchedList
